+++ canvas-design-tools/CHANGELOG.md
@@ -7,6 +7,7 @@ Canvas Design Tools uses [semantic versioning](https://semver.org/).
 * Added support for all previous project locations of _design.config.js_
 * Updated so that typography classes are sorted based on `sorting` config
 * Fixed default config not supporting paragraph indent tokens
+* Fixed design config trying to use Basis paths helper
 
 ## 1.1.0 - 2022-11-16